TWEAKED GAMEPLAY // LEVEL 4
Finished up the fourth level (and lowered the max level amount to 5 because I don't want to overwork myself). At the same time, I changed how dashing works to be more similar to Celeste; the dash now resets when you touch the ground (and your player darkens in color when your dash is used up, changing back to normal once the dash can be used again).