FINAL BUILD
I decided that I'd leave the game like this for the 5/9 deadline. If there is (for some reason) more time after that point I will work on it some more, but I think now would be a good time to just finalize it and get it ready to be submitted.